keepai/services/realtime/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
python-engineio==4.8.0
python-dotenv==1.0.0
eventlet==0.33.3
gunicorn==21.2.0
requests==2.31.0 